Cast to long long to avoid ambiguity in append call

+++ src/mongo/bson/mutable/mutable_bson_builder.cpp
@@ -275,7 +275,7 @@ namespace mutablebson {
             break;
         }
         case NumberLong: {
-            dst->appendNumber(src.fieldName(), src.getLongValue());
+            dst->appendNumber(src.fieldName(), static_cast<long long>(src.getLongValue()));
             break;
         }
         case MaxKey: {
